+++ src/HotChocolate/Core/src/Types.Mutations/Errors/ErrorObjectType.cs
@@ -18,7 +18,7 @@ internal class ErrorObjectType<T> : ObjectType<T>
     {
         // if a user provides his/her own error interface we will not rewrite the message type
         // and the user is responsible for ensuring that type and interface align.
-        if (context.ContextData.ContainsKey(ErrorContextData.ErrorType))
+        if (context.ContextData.ContainsKey(ErrorContextDataKeys.ErrorType))
         {
             return;
         }
